feat(*) Initial Commit
This cl adds the overall basics of the app. When this is finished, the
plan is this will be a lightweight canvas pixel art library, baked with
a bunch of features to enable designers/developers create whatever they
want.

For now all I have built out is the following:

* Drawing circles
* Drawing lines
* Drawing rectangles
* Changing the fill color (circle, rectangles)
* Changing the line color (lines)

Each action gets pushed to an `actionHistory` array, and the canvas
render process is simply a loop of the actionHistory.

In the future I'll support updating previous actions (aka "layers"), and
going backward with undo/redos... as well as a "preview" of what the
user is about to do.

And I'd like to clean up the UI a little, maybe even make a little
axololotl logo or something... maybe :)