feat(Layers) Delete Layers
This CL updates the app to make use of "layers" for deletions. With
this, a user can create several shapes, and delete shapes in the middle.
Each time a shape is created a new layer gets added. If the user hovers
their mouse over the layer, a preview of what they will delete will be
displayed.

I also went back and changed some of the behavior to be moved into
their own functions. I might change the overall architecture of the app
to be more "redux"-ey... but the app hasn't been complex enough to
warrant a framework or flow... but I'll probably make the update in the
future for funsies.