feat(Layers) Move Existing Layer
This CL adds the ability to move existing layers of shapes on the board.
Most of the work needed for this was to add a new enum to handle the
different types of interactions on the layer itself. Right now its just
`SELECT` (for selecting a layer to move) and `DELETE`. I'm employing
a little bit of event delegation for the layers element. I'd imagine the
number of layers that gets created can grow large... so rather than each
element contain a bunch of event handlers for the different parts, it
made more sense to handle that up top in a single event handler.

Fotunately the way this is set up... its not going to be too hard to add
more actions. Adding an `ActionEvent` is simply a matter of creating a
new enum, and handling the enum in the switch-cases for the canvas'
mousedown, mousemove, and mouseup.