feat(Layers) Update layer fill and line colors
This CL adds the ability to change the line and fill colors of previous
layers. With this, I've also updated the UI for the actions (using the
material glyphicons), as well as updating the UI for individual layers.

Instead of dumping the description of the actionEvent in the layer, we
render the shape icon, and color inputs.

I've also fixed an issue with the "MOVE" action. When moving a line,
only the start point of the line was being moved, not the end point.
So.. fixed that.

Might want to start writing unit tets soon. I have a handful of
features I still want to write though... ¯\_(ツ)_/¯