tech(Flux) Introduce a flux architecture
This CL introduces a flux architecture into the code base. I was
strongly considering using redux, but I thought it would be more fun to
play with rxjs some more, so I am trying to re-shift the code to be more
observable/observers based.

An interesting link I saw on this was:
http://rudiyardley.com/redux-single-line-of-code-rxjs/

So yeah, the basic flux architecture is in place. I'm not sure how I'm
going to pull out the app state into the rest of the flows, but it will
be a fun to poke and play with.