fix(CanvasItem) Line movement should not be special
This cl fixes an issue that would start to snowball if left unchecked.
The problem was shapes rendered consisted of the initial x and y
coordinate (`startX`, `startY`), and some type of modifiers.

These modifiers include `width`, `height`, `radius`. The problem was
with the `line` canvasItem. The line used `startX`, `startY`, `endX`,
and `endY`. When using the `MOVE` action, moving lines had a special bit
of logic to update the start and end values in the same way (which
should make sense, otherwise we are just moving the start pivot.)

But that can get messy, especially once I add in a pencil / paint
action. Instead of storing separate x and y end values... I should store
their offset relativev to the startX or startY. This way, if I modify
the startX or startY with the MOVE action, I don't need to worry about
anything else.