tech(Flux) Removed state from app.js
This cl removes all the local state from app.js, and moved it into the
store. This was... a whole lot of shuffling around business logic.

This excercise was ...interesting. Moving everything into actions and a
store ultimately simplified the logic. When performing mousemove or
mouseup from the canvas mousedown, the logic changed from passing around
a generated canvasItem... to passing around the only things that were
available, start and end x and y coordinates.

In the reducer, we then make use of the x and y coordinates and whats
available in the app's state to create the next canvasItem.

I'll be adding more features in the future, but the overall architecture
is there.

Another good next step will be to add tests. I'll do one or the other.

Also gotta make the ui prettier.