tech(Flux) Rotate Action
This CL adds the `ROTATE` action. With this, we can rotate existing
shapes, which is a super helpful tool.

I had to change up how the canvas renders shapes. What I learned in
writing this is that to rotate a shape I need to:

* Rotate rotate the entire canvas
* Draw the shape normally
* Unrotate the canvas

Thats step 1, and with that we have a newly rotated shape. But the
problem is when rotating the canvas, we are rotating about the {0,0}
point (the very top left). Thats going to give super undesirable
results.

So steps become:

* Translate the top left point to our shape
* Rotate rotate the entire canvas
* Draw the shape normally
* Unrotate the canvas
* Translate the top left point back the {0,0} point.

So I tried that out, using the startX and startY coordinates as the spot
to translate to and from intially, and everything worked... with a
slight problem. Rotating from the initial startX and startY coordinates
lead to the shape being rotated from a corner of the shape... which felt
super unnatural.

To fix that, I simply found the centerX and centerY point while
rendering the shape.  The center is different on a per-shape basis.

For lines and rectangles the center is the half-way point between the
start and end coordinates (rectangles using width and height, lines
using endX and endY). Circles, on the other hand, take their center from
the initial startX and startY.

I could have stored the centerX and centerY coordinates as part of the
canvasItem... but because its a derived value, I didn't really want to.
If I manually stored a reference to the center coordinate, then I need
to maintain it throughout the life of that canvasItem. If I move the
canvasItem (which is simply moving the startX and startY coordinates),
then I need to move the center just as much. Trying to maintain a value
that is derived from other values can lead to bugs, which I'd like to
avoid.

If in the future I want to add the ability to rotate based on a movable
pivot point... then I'll need to revisit this and will probably create a
separate centerX and centerY values.