feat(CanvasAction) Brush strokes
This cl adds the brush stroke canvas action. With this, we now have the
ability to free draw a line.

The implementation wasn't that complicated. While the mouse moves across
the canvas, we want to capture the coordinates that the mouse moves
along (using the starting mousedown as the reference point.) We gather
this array of points (path).

Rendering the brush stroke simply involves traversing the path and
moving a line across each point. I've added a throttle to the mousemove
to keep the size of the path manageable.

Rotation was a little hairier. I did a pass through all the points,
grabbing the left, right, top, and bottom-most coordinates, and treated
the center of these coordinates as the center. For the general use-case
this is good enough, but if we play with right triangles we'll see that
the center ends up being the center of the hypotenuse... not the
triangle... whusp. Oh well, I'll go back to fix that in the future some
day.

Oh yeah, this excercise revealed that rendering the canvas is becoming
more and more complicated... so I'll be splitting that into its own
separate component and write some tests on it.

Eventually I'll get to making this app prettier... :)