tech(CanvasRenderer) Creating CanvasRenderer
This cl pulls out the part of app.js that was being used to render the
canvas into its own separate component. With this, I can start testing
out individual parts of how to render canvasItems.

I would have started writing the tests now, except I wasn't sure of the
best way to write spies for mocha, so I punted that to the next cl down
the line.