fix(CanvasItem) Canvas mousemove instead of doc
This cl fixes an issue that was leading to buggy behavior.
The problem was that in implementing the drag-n-drop behavior of the
canvas, I used the mousemove and mouseup of the document for drawing new
items. For normal drag and drop behaviors this makes sense.. but for
drawing on a canvas, it makes a little less sense.

Also that bug. I was grabbing the `offsetX` and `offsetY` from the
mousemove event. The problem was the that `offsetX` and `offsetY` are
relative to the target node. If I move the mouse outside of the canvas,
the target node has a very different offsetX and offsetY. It looked
messy... so I changed things up.

On mousedown of canvas we start the flow, listening to mousemovements on
the canvas. The mousemovement will continue to `takeUntil` we get a
mouseup on the document. So if the user dispatches a mousedown on the
canvas, moves their mouse away, we are only using the last mousemove on
the canvas as our last mousemove event. If the user moves their mouse
back to the canvas, we continue rendering the preview like nothing has
changed. Once the user dispatches mouseup, regardless of if they are
over the canvas or not, we save the canvasItem.