tech(Mobile) Touch Support
This cl adds touch support to this app. I tested the app on my phone and
noticed that nothing canvas related was working. The root cause was
because I was adding event handlers to `mousedown`, `mousemove`,
`mouseup`... and not `touchstart`, `touchmove`, and `touchend`.

It was incredibly easy to add touch support. I just needed to create
observables from the additional events, and merge them.

The only thing I needed to watch out for is the x/y coordinates of mouse
events differ from the x/y coordinates of touch. The simple fix was to
`map` the events to return `startX`, `endX`, `localEndX`, and
`localEndY` independently.