design(Mobile) Better responsive ui
This cl updates the layout of the app to work better for mobile web. All
the actions appear on 1 line at the top of the page. In mobile views,
the layout is changed from column-based to row-based, with the actions
appearing to the left of the canvas.

The ui is far from perfect. The mobile layout is still hiding the layers
from view. I'll get back to fixing that later.