tech(CanvasAction) Copy Action
This CL adds the "copy" action to the list of available actions. After
adding a shape, the copy action will create a shallow duplicate of the
selected canvasItem.

I changed up the logic a bit too for this. Previously `id` came from a
`scan` found in the canvas' mousedown. But... the problem with this
approach is when making a copy, I need to assign a new id. If the ids
were coming from the mousedown event... then I wouldn't know how to pick
the "next" id.

One thing I did notice from implementing copy is there is a bug with the
layers. I've been maintaining the order of the CanvasItems, but when
rendering the layers I was simply prepending the layer to the list...
which doesn't work as expected for copying layers.

Thats a bug fix for a later time.