feat(CanvasAction) Adds linewidth
This CL adds linewidth as a modifyable value for `ctx.stroke`. Its not a
big change, but it does provide some additional customization.

Following this I might want to take some time to make the customization
for line/fill color and line width easier. I'm currently not a fan of
having to bring my mouse back to an input to make those updates... so I
might be adding some keyboard shortcuts to create custom color
pickers... hopefully something intuitive but pretty.