fix(CanvasAction) Fix duplicate bug
This cl fixes a bug when there layers where appearing in a different
order than the canvasItemList. The layers should appear in reverse order
of the canvasItemList.

Previously we were just prepending new canvasItem layers to the front of
the nodelist. But, if we duplicated a layer, the duplicate would appear
right next to the source in the canvasItemList.. but the layers were
continuing to promote the duplicated item to the front.

A shoddy illustration:

With bug:

```
canvasItem = [1, 2, 3];
layers = ['3', '2', '1'];

duplucate(1);

canvasItem = [1, 4, 2, 3];
layers = ['4', '3', '2', '1'];
```

With this fix:

```
canvasItem = [1, 2, 3];
layers = ['3', '2', '1'];

duplucate(1);

canvasItem = [1, 4, 2, 3];
layers = ['3', '2', '4', '1'];
```

This is a minor fix.  ... And to be honest its not a complete fix
either, since I'm going to want to handle reordering layers in the near
future.... :\