test(CanvasRender) renderCanvasItem
This cl introduces tests for the `renderCanvasItem` method. And with
this, we have near complete coverage of the different use-cases of
`canvas-render`.

Back to features soons maybe. Ideally future cls will include the tests
as I write them, but sometimes its hard to be motivated to formally
write tests everywhere with pet projects. Ultimately it will end up
hurting me down the line if I skip tests, so... heres to hoping.