fix(Mobile) Dragging window while drawing
This cl maybe fixes an issue with the mobile experience. When dragging
my finger across the canvas to draw a picture, the page would scroll
about.

I've made 2 changes:
1. useCapture
2. preventDefault.

Using the capture phase is probably unnecessary, but it sort of stood to
reaosn that if I am dragging my mouse or finger around on a canvas I
want to handle it downward rather than upward.

`preventDefault` might be my best bet in seeing if the page stops
dragging on the mobile experience, but I won't know for sure until I
test this.

I *could* have gotten away with setting up a local server on my laptop
and using the iphone simulator to test this.. but I had a mix of lazy
and being skeptical that the simulator will produce the same result that
I am expecting when I use a physical devices (especially with iphone...)