feat(CanvasAction) Smooth Draw x/y coordinates
This cl changes the behavior of drawing to "smooth" the mousemovement if
the `shiftKey` is being held. Its a small feature, but it makes drawing
a lot easier if I want straight lines.

This cl also fixes a bug with shift too. If I move my mouse around while
holding `shift` things start to get highlighted and the page scrolls
around... which is definitely a bug.